fix: sidebar agora está fixa e ocupa a tela inteira
